test: split useRecommendations suites
